+++ R/bfs_get_catalog.R
@@ -1,61 +1,29 @@
-#' Get the Swiss Federal Statistical Office data catalog in a given language
+#' Get the BFS data catalog
 #'
-#' This function scraps the RSS Feed of the Swiss Federal Statistical Office data catalog.
+#' This function scraps a given RSS Feed of the Swiss Federal Statistical Office.
 #'
 #' @param language character The language of a BFS catalog.
+#' @param type character A BFS catalog
 #'
 #' @return A data frame
 #'
 #' @importFrom tidyRSS tidyfeed
 #' @importFrom janitor clean_names
-#' @importFrom dplyr select mutate
 #'
-#' @seealso \code{\link{bfs_get_metadata}}
+#' @seealso \code{\link{bfs_get_data}}
 #'
 #' @examples
-#' \donttest{bfs_get_catalog(language = "de")}
+#' \donttest{bfs_get_catalog(language = "de", type = "data")}
 #'
 #' @export
-
-bfs_get_catalog <- function(language = c("de", "fr", "it", "en")) {
-  if (missing(language))
-    stop("must choose a language, either 'de', 'fr', 'it' or 'en'", call. = FALSE)
+bfs_get_catalog <- function(language = c("de", "fr", "it", "en"), type = c("data", "tables")) {
   
-  language <-
-    match.arg(arg = language, choices = c("de", "fr", "it", "en"))
+  if (missing(language)) stop("must choose a language, either 'de', 'fr', 'it' or 'en'", call. = FALSE)
+  language <- match.arg(arg = language, choices = c("de", "fr", "it", "en"))
+  if (missing(type)) stop("must choose a type, either 'data' or 'tables'", call. = FALSE)
+  type <- match.arg(arg = type, choices = c("data", "tables"))
+    
+  if(type == "data") { BFS::bfs_get_catalog_data(language = language) }
+  if(type == "tables") { BFS::bfs_get_catalog_tables(language = language) }
   
-  feed <-
-    paste0(
-      "https://www.bfs.admin.ch/bfs/",
-      language,
-      "/home/statistiken/kataloge-datenbanken/daten/_jcr_content/par/ws_catalog.rss.xml?skipLimit=true"
-    )
-  
-  df <- tidyRSS::tidyfeed(feed = feed)
-  
-  colnames(df) <- gsub('feed_', '', colnames(df))
-  colnames(df) <- gsub('item_', '', colnames(df))
-  
-  df <- janitor::clean_names(df, "small_camel")
-  
-  base_url_bfs <-
-    paste0(
-      "https://www.bfs.admin.ch/content/bfs/",
-      language,
-      "/home/statistiken/kataloge-datenbanken/daten.assetdetail."
-    )
-  base_url_px <- "https://www.bfs.admin.ch/bfsstatic/dam/assets/"
-  
-  df <- df %>%
-    dplyr::select(title = title_2,
-                  url_bfs = link_2,
-                  published = pubDate_2) %>%
-    dplyr::mutate(language = language) %>%
-    dplyr::mutate(
-      url_px = gsub(base_url_bfs, base_url_px, url_bfs),
-      url_px = gsub(".html$", "/master", url_px)
-    ) %>%
-    dplyr::select(title, language, published, url_bfs, url_px)
-  
-  return(df)
 }